Split leases schema into different files
